+++ react-tenant-selector/src/components/LoginOptions/LoginWithADFS.tsx
@@ -2,18 +2,14 @@ import React from 'react';
 import { Button } from '@cognite/cogs.js';
 import { CogniteAuth, saveFlow } from '@cognite/auth-utils';
 
-const LoginWithAzureAD = ({
-  cluster,
-  authClient,
-}: {
-  cluster: string;
+interface Props {
   authClient?: CogniteAuth;
-}) => {
+}
+const LoginWithADFS: React.FC<Props> = ({ authClient }: Props) => {
   const handleClick = () => {
     if (authClient) {
-      // console.log('Starting AZURE_AD login flow');
-      saveFlow('AZURE_AD');
-      authClient.login('AZURE_AD', { cluster });
+      saveFlow('ADFS');
+      authClient.login('ADFS');
     }
   };
 
@@ -24,9 +20,9 @@ const LoginWithAzureAD = ({
       type="secondary"
       onClick={handleClick}
     >
-      Login with Microsoft Azure
+      Login with ADFS
     </Button>
   );
 };
 
-export default LoginWithAzureAD;
+export default LoginWithADFS;