+++ react-tenant-selector/src/components/LoginOptions/index.ts
@@ -1,4 +1,4 @@
 export { default as LoginWithAzure } from './LoginWithAzure';
-export { default as LoginWithAzureAD } from './LoginWithAzureAD';
+export { default as LoginWithADFS } from './LoginWithADFS';
 export { default as LoginWithCognite } from './LoginWithCognite';
 export { default as LoginOrWrapper } from './LoginOrWrapper';