+++ react-tenant-selector/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as CardContainerHeader } from './CardContainerHeader';
-export { default as CardFooterError } from './CardFooterError';
+export { default as Error, ErrorExpandable } from './Errors';
 export { default as ClusterSelector } from './ClusterSelector';
 export { default as CogniteFlow } from './CogniteFlow';
 export { default as LoginTip } from './LoginTip';